pyproject.toml
[tool.poetry]
name = "nonebot_plugin_memes"
version = "0.8.0"
description = "Nonebot2 plugin for making memes"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-memes"
repository = "https://github.com/noneplugin/nonebot-plugin-memes"

[tool.poetry.dependencies]
python = "^3.9"
nonebot2 = "^2.3.0"
nonebot-plugin-alconna = ">=0.46.3,<1.0.0"
nonebot-plugin-uninfo = ">=0.6.1,<1.0.0"
nonebot-plugin-localstore = ">=0.6.0,<1.0.0"
nonebot-plugin-orm = ">=0.7.0,<1.0.0"
meme-generator = "^0.2.0"
pyyaml = "^6.0"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
nonebot2 = { extras = ["fastapi", "httpx"], version = "^2.3.0" }
nonebot-plugin-orm = { extras = ["default"], version = ">=0.7.0,<1.0.0" }
nonebot-adapter-onebot = "^2.2.0"

[tool.nonebot]
plugins = ["nonebot_plugin_memes"]

[tool.pyright]
pythonVersion = "3.9"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
typeCheckingMode = "basic"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["E402", "C901", "UP037"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
